rename: AlaramController 위치 및 dto 위치변경
