+++ src/main/java/com/grepp/spring/app/model/alarm/dto/response/AlarmListResponse.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.controller.api.alarm.payload;
+package com.grepp.spring.app.model.alarm.dto.response;
 
 import com.grepp.spring.app.model.alarm.entity.Alarm;
 import com.grepp.spring.app.model.alarm.entity.AlarmRecipient;
